% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{automata_select}
\alias{automata_select}
\title{Standardised Input Selector for Automata Filtering}
\usage{
automata_select(input_id, label = "Include automata (e.g. web crawlers)")
}
\arguments{
\item{input_id}{ID}

\item{label}{Label}
}
\description{
Creates a checkbox for including/excluding automata.
}
\seealso{
Other inputs: 
\code{\link{smooth_select}()},
\code{\link{timeframe_daterange}()},
\code{\link{timeframe_select}()}
}
\concept{inputs}
