% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Capitalize First Letter Of Every Word}
\source{
\url{https://stackoverflow.com/questions/6364783/capitalize-the-first-letter-of-both-words-in-a-two-word-string}
}
\usage{
capitalize_first_letter(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Capitalizes the first letter of every word.
}
\details{
This function is made available under CC-BY-SA 3.0
}
\author{
\href{https://stackoverflow.com/users/602276/andrie}{Andrie de Vries}
}
