% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{cbind_fill}
\alias{cbind_fill}
\title{Safely Combine R Objects of Variying Lengths by Columns}
\usage{
cbind_fill(...)
}
\arguments{
\item{...}{Vectors or matrices.}
}
\value{
A matrix with NAs wherever needed.
}
\description{
Take a sequence of vector, matrix or data-frame arguments and
combine by columns or rows, respectively.
}
\references{
\url{http://r.789695.n4.nabble.com/How-to-join-matrices-of-different-row-length-from-a-list-td3177212.html}
}
