% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_notify.R
\name{check_notify}
\alias{check_notify}
\alias{check_past_week}
\alias{check_yesterday}
\title{Check 'n' Notify}
\usage{
check_yesterday(dataset, label)

check_past_week(dataset, label)
}
\arguments{
\item{dataset}{A data.frame with a 'date' column.}

\item{label}{The label to use in the notification.}
}
\description{
Check a dataset for missing data from the day before and the
 past week, and create a \code{notificationItem} if missing data has been
 detected.
}
\seealso{
Other Shiny Dashboarding: \code{\link{cond_color}},
  \code{\link{cond_icon}}, \code{\link{na_box}},
  \code{\link{time_frame_range}}
}

