% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dygraphs.R
\name{cond_color}
\alias{cond_color}
\title{Select a Colour Conditionally}
\usage{
cond_color(condition, true_color = "green")
}
\arguments{
\item{condition}{a condition to be evaluated to produce a single TRUE or FALSE value}

\item{true_color}{the colour used to represent a TRUE result. Green by default.}
}
\description{
select a colour based on the true/false nature of a condition.
Uses green as the "true" colour by default, "red" as false, and
}

