% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphs.R
\docType{data}
\name{custom_axis_formatter}
\alias{custom_axis_formatter}
\title{Custom Axis Formatter}
\format{An object of class \code{character} of length 1.}
\usage{
custom_axis_formatter
}
\description{
This is a JS date formatter to be used in \code{dyAxis} when
it's desired to have the dates on the axis look like "Monday (8/13)", in
other words: "Day of the week (MM/DD)"
}
\examples{
\dontrun{
polloi::make_dygraph(data) \%>\%
 dyAxis("x", axisLabelFormatter = polloi::custom_axis_formatter)
}
}
\keyword{datasets}
