% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{data_select}
\alias{data_select}
\title{Conditionally Select A Dataset}
\usage{
data_select(test, yes_set, no_set)
}
\arguments{
\item{test}{an object which can be coerced to logical mode.}

\item{yes_set}{the dataset to return if \code{test} is true.}

\item{no_set}{the dataset to return if \code{test} is false.}
}
\value{
The appropriate object.
}
\description{
Like \code{ifelse}, but not awful and usable with
 \code{data.frames}.
}

