% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_langproj}
\alias{get_langproj}
\title{Wikimedia Language-Project Tuples}
\usage{
get_langproj()
}
\description{
Returns a dataset containing all the projects and all possible
  language-project combinations for multilingual projects.
}
\details{
We use a Cartesian product here because we cannot trust the matrix
  of languages and projects to be up-to-date, because it is not usually
  updated when a new wiki is deployed. The product method ensures that all
  the multilingual wikis are correctly captured (provided we have the
  latest prefixes) by \code{\link{parse_wikiid}}.
}
\seealso{
\code{\link{get_prefixes}}, \code{\link{get_projects}}
}

