% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_projects}
\alias{get_projects}
\title{Wikimedia Projects}
\format{A data frame with 280-some rows and 3 variables:
\describe{
\item{wikiid}{The wiki ID used internally.}
\item{project}{Name of the Wikimedia project.}
\item{multilingual}{A boolean indicator whether the project has language
subdomains.}
}}
\source{
\url{https://meta.wikimedia.org/wiki/Table_of_Wikimedia_projects}
}
\usage{
get_projects()
}
\description{
Returns a dataset containing almost 40 Wikimedia Foundation
projects.
}
\seealso{
\link{update_projects}, \link{get_prefixes}, \link{get_langproj}
}
