% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_us_state}
\alias{get_us_state}
\title{U.S. States and Regions}
\format{A data frame with 56 rows and 3 variables:
\describe{
\item{abb}{The abbreviations of U.S. states and territories.}
\item{region}{The regions of U.S. states and territories. See
https://phabricator.wikimedia.org/T136257#2399411.}
\item{state}{The names of U.S. states and territories.}
}}
\source{
\code{ISO_3166_1} from package \code{ISOcodes}; \code{state.name}, \code{state.abb} and
\code{state.region} from package \code{datasets}; see
\url{https://phabricator.wikimedia.org/T136257#2399411} for U.S. regions.
}
\usage{
get_us_state()
}
\description{
Returns a dataset containing all U.S. states' and territories'
names, abbreviations and regions.
}
