% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{half}
\alias{half}
\title{Sample Half an Object}
\usage{
half(x, top = TRUE)
}
\arguments{
\item{x}{the object to sample from}

\item{top}{whether it should be the top (TRUE) or bottom (FALSE) half. Set
to TRUE by default.}
}
\description{
easily sample the top or bottom half of an object.
}
