% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maths.R
\name{mad}
\alias{mad}
\title{Median Absolute Deviation}
\usage{
mad(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
computes the Median Absolute Deviation, or MAD. It's like standard deviations
only it doesn't suck terribly for any dataset we'd actually encounter outside a classroom.
}

