% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphs.R
\name{make_dygraph}
\alias{make_dygraph}
\title{Construct a Standard Wikimedia Discovery Dygraph}
\usage{
make_dygraph(data, xlab, ylab, title, legend_name = NULL, use_si = TRUE,
  expr = NULL, group = NULL, ...)
}
\arguments{
\item{data}{a data.frame reformatted to be XTS-able.}

\item{xlab}{the label to use for the dygraph's x-axis.}

\item{ylab}{the label to use for the dygraph's y-axis.}

\item{title}{the title to use on the dygraph.}

\item{legend_name}{a custom name for the variable in the event that \code{is_single} is TRUE.
NULL by default (in which case the variable will be named "events").}

\item{use_si}{whether to use si labelling (1000 becomes 1K). TRUE by default.}

\item{expr}{an optional expression to evaluate prior to building the dygraph.
We use this in (for example) reactive graphing.}

\item{group}{Group to associate this plot with. The x-axis zoom level of
plots within a group is automatically synchronized.}

\item{...}{Additional parameters to pass on to \code{dyOptions}.}
}
\description{
Construct a dygraph using the custom formatting Wikimedia dashboards use
as standard. This is nothing special - a standard dygraph with a bit of custom CSS
shipped with the package - but it's surrounded by code that allows the function to
turn all our usual data.frame formats into XTS objects.
}

