% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{na_box}
\alias{na_box}
\title{Bad Data Value Box}
\usage{
na_box(subtitle)
}
\arguments{
\item{subtitle}{Subtitle text.}
}
\description{
This is used in cases where something has gone wrong with the
 data and a value could not be calculated. This ensures the "bad data" value
 boxes have a consistent look.
}
\seealso{
Other Shiny Dashboarding: \code{\link{check_notify}},
  \code{\link{cond_color}}, \code{\link{cond_icon}},
  \code{\link{time_frame_range}}
}

