% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_wikiid}
\alias{parse_wikiid}
\title{Parse Wikiid into Language-Project}
\usage{
parse_wikiid(x)
}
\arguments{
\item{x}{A character vector of wikiid's.}
}
\value{
A data frame with columns 'language' and 'project'.
}
\description{
Uses the internal index of languages, prefixes, wikis, and
 projects to parse a wikiid into a language-project tuple.
}

