% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maths.R
\name{percent_change}
\alias{percent_change}
\title{Calculate a Percentage Change}
\usage{
percent_change(x, y = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{(optionally) an additional numeric vector. If y is not provided, each
value in x will be compared to the subsequent value - if y is, each x value will
be compared to the equivalent y value.}
}
\description{
calculates a delta between X or Y (or sequential X values)
expressed as a percentage.
}

