% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{portal_regions}
\alias{portal_regions}
\title{Countries and Regions with Traffic to Wikipedia.org}
\format{\code{portal_regions} is a character vector containing about 230
country/region names with traffic to Wikipedia.org (portal).}
\source{
\url{https://analytics.wikimedia.org/datasets/discovery/metrics/portal/all_country_data.tsv}
}
\usage{
portal_regions
}
\seealso{
\link{update_portal_regions}
}
\keyword{datasets}
