% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading.R
\name{read_dataset}
\alias{read_dataset}
\title{Read Datasets from datasets.wikimedia}
\usage{
read_dataset(path)
}
\arguments{
\item{path}{the path to the file, starting from datasets.wikimedia.org/aggregate-datasets/}
}
\description{
Grabs datasets from the public directories at
\href{https://datasets.wikimedia.org}{datasets.wikimedia.org} for use in dashboarding.
}

