% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{reorder_columns}
\alias{reorder_columns}
\title{Reorder Columns Based On Last Observed Value}
\usage{
reorder_columns(x, decreasing = TRUE)
}
\arguments{
\item{x}{a \code{data.frame} where the first column is the date and the rest are
columns to be reordered}

\item{decreasing}{which way to sort; \code{TRUE} by default}
}
\description{
When visualizing multiple series in a dygraph, it can be
helpful to have the series be in the same order in the legend.
}
