% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphs.R
\name{smart_palette}
\alias{smart_palette}
\title{Intelligently choose the correct color palette}
\usage{
smart_palette(n)
}
\arguments{
\item{n}{Number of colors to return (number of categories).}
}
\value{
A character vector of \code{n} colors.
}
\description{
Picks between 3 different color palettes based on number of
 categories to visualize. For <13 categories, uses Set1, Paired, & Set3
 from RColorBrewer. For More than 12 categories, builds a HCL color
 space-based palette.
}

