% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{smooth_select}
\alias{smooth_select}
\title{Standardised Input Selector for Smoothing}
\usage{
smooth_select(input_id, label = "Smoothing")
}
\arguments{
\item{input_id}{ID}

\item{label}{Label}
}
\seealso{
Other inputs: \code{\link{automata_select}},
  \code{\link{timeframe_daterange}},
  \code{\link{timeframe_select}}
}
