% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smooth_switch}
\alias{smooth_switch}
\title{Switch between global and local smoothing}
\usage{
smooth_switch(global, local)
}
\arguments{
\item{global}{the input variable for global smoothing.}

\item{local}{the input variable for local smoothing.}
}
\description{
We use a lot of smoothing in our reactive graphs, and tend to
 offer both global (entire dashboard) and local (tab-specific) smoothing
 options. \code{smooth_switch} is a simple function to abstract away the
 logic behind determining whether the global or local option should be
 relied on.
}

