% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{time_frame_range}
\alias{time_frame_range}
\title{Get The Time Range}
\usage{
time_frame_range(
  input_local_timeframe,
  input_local_daterange,
  input_global_timeframe,
  input_global_daterange
)
}
\arguments{
\item{input_local_timeframe}{The value of the input corresponding to the
local timeframe selector.}

\item{input_local_daterange}{The value of the input corresponding to the
local date range selector.}

\item{input_global_timeframe}{The value of the input corresponding to the
global timeframe selector.}

\item{input_global_daterange}{The value of the input corresponding to the
global date range selector.}
}
\value{
A Date vector of length 2
}
\description{
This figures out the date range based on the four different
inputs: local & global timeframe selections and local & global date range
selections. It is used for \link{subset_by_date_range}.
}
\examples{
\dontrun{
time_frame_range(
  input$timeframe, input$daterange,
  input$timeframe_global, input$daterange_global
)
}
}
\seealso{
Other Shiny Dashboarding: 
\code{\link{check_notify}},
\code{\link{cond_color}()},
\code{\link{cond_icon}()},
\code{\link{na_box}()}
}
\concept{Shiny Dashboarding}
