% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{timeframe_daterange}
\alias{timeframe_daterange}
\title{Standardized Date Range Selector for Time Frame}
\usage{
timeframe_daterange(select_input_id, label = "Custom Date Range")
}
\arguments{
\item{select_input_id}{The ID you used for the corresponding
\link{timeframe_select}}

\item{label}{Label}
}
\value{
A \link[shiny:conditionalPanel]{shiny::conditionalPanel} containing a \link[shiny:dateRangeInput]{shiny::dateRangeInput}
}
\description{
Creates a date ranger selector.
}
\seealso{
Other inputs: 
\code{\link{automata_select}()},
\code{\link{smooth_select}()},
\code{\link{timeframe_select}()}
}
\concept{inputs}
