% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{timeframe_select}
\alias{timeframe_select}
\title{Standardized Drop-down Selector for Time Frame}
\usage{
timeframe_select(input_id, label = "Time Frame")
}
\arguments{
\item{input_id}{ID}

\item{label}{Label}
}
\value{
A \code{selectInput}
}
\seealso{
timeframe_daterange "Shiny Dashboarding"

Other inputs: \code{\link{smooth_select}},
  \code{\link{timeframe_daterange}}
}

