% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_active_wikis}
\alias{update_active_wikis}
\title{Update internal index of active wikis}
\usage{
update_active_wikis(dev)
}
\arguments{
\item{dev}{Logical flag that controls whether fetched prefix list is saved
to the \strong{data/} directory (\code{dev = TRUE}) or to where \emph{polloi} is
installed (default)}
}
\description{
Fetches & processes site matrix to get a list of active wikis.
}
