% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_portal_regions}
\alias{update_portal_regions}
\title{Update Country and Region Names with Traffic to Wikipedia.org}
\usage{
update_portal_regions(dev = FALSE)
}
\arguments{
\item{dev}{Logical flag that controls whether fetched country list is saved
to the \strong{data/} directory (\code{dev = TRUE}) or to where \emph{polloi} is
installed (default)}
}
\description{
Get unique country and region names from the \strong{country} column of
\url{https://analytics.wikimedia.org/datasets/discovery/metrics/portal/all_country_data.tsv}.
}
