% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_projects}
\alias{update_projects}
\title{Update internal index of wiki projects}
\usage{
update_projects()
}
\description{
Updates the internal list of Wikimedia projects from the
package repository to avoid having to reinstall the package from source.
}
