% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wdqs_usage}
\alias{wdqs_usage}
\title{Wikidata Query Service usage dataset}
\format{A \code{data.frame} with over 600 rows and 4 columns:
\describe{
\item{date}{\code{Date}}
\item{homepage}{Web requests made to query.wikidata.org/}
\item{LDF endpoint}{Requests made to query.wikidata.org/bigdata/ldf}
\item{SPARQL endpoint}{Requests made to \href{https://www.wikidata.org/wiki/Wikidata:SPARQL_query_service}{WDQS SPARQL endpoint}} # nolint
}}
\source{
\url{https://analytics.wikimedia.org/datasets/discovery/metrics/wdqs/basic_usage.tsv}
}
\usage{
wdqs_usage
}
\description{
WDQS usage dataset is used for unit testing and examples.
}
\seealso{
\link{get_sample_data}
}
\keyword{datasets}
