% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{compress}
\alias{compress}
\title{Convert Numeric Values to use SI suffixes}
\usage{
compress(x, round_by = 2)
}
\arguments{
\item{x}{a vector of numeric or integer values}

\item{round_by}{how many digits to round the resulting numbers by.}
}
\description{
takes a numeric vector (1200, 1300, 1400) and converts it to
use SI suffixes (1.2K, 1.3K, 1.4k)
}

