% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dygraphs.R
\name{cond_icon}
\alias{cond_icon}
\title{Select an appropriate directional icon}
\usage{
cond_icon(condition, true_direction = "up")
}
\arguments{
\item{condition}{a condition to be evaluated to produce a single TRUE/FALSE}

\item{true_direction}{which direction represents a positive change. "up" by
default.}
}
\description{
allows you to select an appropriate directional icon for a change
in condition.
}

