% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_dataset}
\alias{read_dataset}
\title{Read Datasets from datasets.wikimedia}
\usage{
read_dataset(path, force_uncached = TRUE, ...)
}
\arguments{
\item{path}{the path to the file, starting from https://analytics.wikimedia.org/datasets/}

\item{force_uncached}{If TRUE, appends a \code{?ts=...} to URL to force an uncached version}

\item{...}{Additional arguments to be passed to \link[readr:read_delim]{readr::read_delim}}
}
\description{
Grabs datasets from the public directories at
\href{https://analytics.wikimedia.org/datasets/}{WMF Analytics' Datasets} for
use in dashboarding.
}
