% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{smoother}
\alias{smoother}
\title{Dynamically Smooth Data}
\usage{
smoother(dataset, smooth_level = "day", rename = TRUE)
}
\arguments{
\item{dataset}{an untidy, dygraph-appropriate data.frame}

\item{smooth_level}{the level of smoothing. Options are "day", "week",
"month", and "gam"}

\item{rename}{whether to rename the fields once smoothed. TRUE by default.}
}
\description{
Takes an untidy (read: dygraph-appropriate) dataset and adds
columns for each variable consisting of the smoothed, averaged mean.
}
