% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_prefixes}
\alias{update_prefixes}
\title{Update internal index of language prefixes.}
\usage{
update_prefixes()
}
\description{
Scrapes the table off
 \url{https://meta.wikimedia.org/wiki/Table_of_Wikimedia_projects} and stores the first
 three columns containing the language, local language, and the prefix code.
}

